style: satisfy clippy -D warnings and fix cross-platform builds
Prepare the workspace so the CI build/lint jobs pass on every host/target.

- Apply `cargo clippy --fix` plus targeted `#[allow]`s to satisfy `-D warnings`
  across the crates and the CLI, and add `clippy.toml` thresholds.
- cfg-gate platform-specific code (apple / `cfg(unix)`) and add a non-apple
  `generate_swift_bridge` stub so the CLI, runner, and lib compile on Linux,
  Windows, and macOS hosts and their cross targets.
- Drop a stale trailing arg in the runner's createPageInstance call.

+++ crates/lingxia-platform/src/desktop/image.rs
@@ -128,7 +128,7 @@ pub fn compress_image_desktop(request: &CompressImageRequest) -> Result<PathBuf,
     let output_file = File::create(&request.output_path)
         .map_err(|e| PlatformError::Platform(format!("Failed to create output file: {}", e)))?;
 
-    let quality = request.quality.clamp(1, 100) as u8;
+    let quality = request.quality.clamp(1, 100);
     let mut encoder = JpegEncoder::new_with_quality(output_file, quality);
 
     encoder
